Introduce a way to have specific key/value names
This is an alternative solution for PR #47.

By using this approach we don’t have to make any changes to the raw
output generated by the `listApis` API call (which is saved in the
`v43.go` and `v44.go` files).

So when we now want to generate code based on a new version, we only
have to paste the output of the `listApis` call info a new `vxx.go`
file and were good to go, without having to manually alter the file.

Next to all that’ it simple feels a little cleaner and more robust
:wink: