Change synchronous boolean response to bool
The `success` fields in newer versions of CloudStack will use a `bool` value instead of a `string`, matching the behavious of asynchronious responses.

To make sure this change is backwards compatible the related types now have a custom `UnmarshalJSON` method that unmarshals the data (string or bool) into the new bool field.