Avoid embedded structs by generating new struct types
This change makes it easier to manually create instances of generated
types. Instead of embedding the structs the generator now creates a new
struct for each embedded object. This also allows easier testing in 3rd
party code using this lib since it's simpler to create.

The downside is that this is a breaking change, but only for people
manually creating instances of embedded structs and I think they will
welcome the change. An error will be raised during compile time which
also makes this change safe.